[charts] Support range-bar series in WebGL bar renderer
Extracts the bar WebGL data assembly into a shared
useWebGLBarLikePlotData hook parametrised by highlight type and
full-rounded-corner mode. Adds a RangeBarWebGLPlot that reuses the
BarWebGLProgram and shaders, and mounts it inside ChartsWebGLLayer
when renderer="webgl" so range-bar series render on the GPU
alongside standard bars.
