+++ packages/x-charts-premium/src/BarChartPremium/webgl/BarWebGLPlot.test.tsx
@@ -59,4 +59,51 @@ describe('<BarWebGLPlot />', () => {
       ),
     ).not.to.throw();
   });
+
+  it.skipIf(isJSDOM)('should render range bar series with the webgl renderer', () => {
+    expect(() =>
+      render(
+        <BarChartPremium
+          width={400}
+          height={300}
+          renderer="webgl"
+          series={[
+            {
+              type: 'rangeBar',
+              data: [
+                [1, 4],
+                [2, 5],
+                [3, 6],
+              ],
+            },
+          ]}
+          xAxis={[{ scaleType: 'band', data: ['a', 'b', 'c'] }]}
+        />,
+      ),
+    ).not.to.throw();
+  });
+
+  it.skipIf(isJSDOM)('should render mixed bar and range bar series with the webgl renderer', () => {
+    expect(() =>
+      render(
+        <BarChartPremium
+          width={400}
+          height={300}
+          renderer="webgl"
+          series={[
+            { data: [1, 2, 3] },
+            {
+              type: 'rangeBar',
+              data: [
+                [2, 4],
+                [3, 5],
+                [1, 3],
+              ],
+            },
+          ]}
+          xAxis={[{ scaleType: 'band', data: ['a', 'b', 'c'] }]}
+        />,
+      ),
+    ).not.to.throw();
+  });
 });
